sharebite_backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg[binary]>=3.2.0
django-cors-headers==4.3.1
python-decouple==3.8
twilio==8.11.1
Pillow
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
geopy==2.4.1
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
